Provides a better integration with sbt-org-policies
