+++ docs/src/main/tut/index.md
@@ -37,4 +37,3 @@ Additionally, it depends on other useful libraries and plugins like:
 * [sbt-scalafmt](https://github.com/olafurpg/scalafmt)
 * [sbt-pgp](https://github.com/sbt/sbt-pgp)
 * [sbt-header](https://github.com/sbt/sbt-header)
-* [sbt-native-packager](https://github.com/sbt/sbt-native-packager)
