+++ src/main/scala/microsites/microsites.scala
@@ -14,124 +14,122 @@
  * limitations under the License.
  */
 
+package microsites
+
 import net.jcazevedo.moultingyaml._
 import sbt._
 
 import scala.language.{postfixOps, reflectiveCalls}
 
-package object microsites {
+case class KazariDependency(
+    groupId: String,
+    artifactId: String,
+    scalaVersion: String,
+    version: String)
 
-  case class KazariDependency(
-      groupId: String,
-      artifactId: String,
-      scalaVersion: String,
-      version: String)
+case class KazariSettings(
+    micrositeKazariEvaluatorUrl: String,
+    micrositeKazariEvaluatorToken: String,
+    micrositeKazariGithubToken: String,
+    micrositeKazariCodeMirrorTheme: String,
+    micrositeKazariDependencies: Seq[KazariDependency],
+    micrositeKazariResolvers: Seq[String])
 
-  case class KazariSettings(
-      micrositeKazariEvaluatorUrl: String,
-      micrositeKazariEvaluatorToken: String,
-      micrositeKazariGithubToken: String,
-      micrositeKazariCodeMirrorTheme: String,
-      micrositeKazariDependencies: Seq[KazariDependency],
-      micrositeKazariResolvers: Seq[String])
+case class MicrositeIdentitySettings(
+    name: String,
+    description: String,
+    author: String,
+    homepage: String,
+    twitter: String)
 
-  case class MicrositeIdentitySettings(
-      name: String,
-      description: String,
-      author: String,
-      homepage: String,
-      twitter: String)
+case class MicrositeFileLocations(
+    micrositeImgDirectory: File,
+    micrositeCssDirectory: File,
+    micrositeJsDirectory: File,
+    micrositeCDNDirectives: CdnDirectives,
+    micrositeExternalLayoutsDirectory: File,
+    micrositeExternalIncludesDirectory: File,
+    micrositeDataDirectory: File,
+    micrositeExtraMdFiles: Map[File, ExtraMdFileConfig])
 
-  case class MicrositeFileLocations(
-      micrositeImgDirectory: File,
-      micrositeCssDirectory: File,
-      micrositeJsDirectory: File,
-      micrositeCDNDirectives: CdnDirectives,
-      micrositeExternalLayoutsDirectory: File,
-      micrositeExternalIncludesDirectory: File,
-      micrositeDataDirectory: File,
-      micrositeExtraMdFiles: Map[File, ExtraMdFileConfig])
+case class MicrositeGitSettings(
+    githubOwner: String,
+    githubRepo: String,
+    gitHostingService: MicrositeKeys.GitHostingService,
+    gitHostingUrl: String)
 
-  case class MicrositeGitSettings(
-      githubOwner: String,
-      githubRepo: String,
-      gitHostingService: MicrositeKeys.GitHostingService,
-      gitHostingUrl: String)
+case class MicrositeUrlSettings(micrositeBaseUrl: String, micrositeDocumentationUrl: String)
 
-  case class MicrositeUrlSettings(micrositeBaseUrl: String, micrositeDocumentationUrl: String)
+case class MicrositeFavicon(filename: String, sizeDescription: String)
 
-  case class MicrositeFavicon(filename: String, sizeDescription: String)
+case class MicrositeVisualSettings(
+    highlightTheme: String,
+    palette: Map[String, String],
+    favicons: Seq[MicrositeFavicon])
 
-  case class MicrositeVisualSettings(
-      highlightTheme: String,
-      palette: Map[String, String],
-      favicons: Seq[MicrositeFavicon])
+case class MicrositeSettings(
+    identity: MicrositeIdentitySettings,
+    visualSettings: MicrositeVisualSettings,
+    configYaml: ConfigYml,
+    fileLocations: MicrositeFileLocations,
+    urlSettings: MicrositeUrlSettings,
+    gitSettings: MicrositeGitSettings,
+    micrositeKazariSettings: KazariSettings) {
 
-  case class MicrositeSettings(
-      identity: MicrositeIdentitySettings,
-      visualSettings: MicrositeVisualSettings,
-      configYaml: ConfigYml,
-      fileLocations: MicrositeFileLocations,
-      urlSettings: MicrositeUrlSettings,
-      gitSettings: MicrositeGitSettings,
-      micrositeKazariSettings: KazariSettings) {
-
-    def gitSiteUrl: String = {
-      gitSettings.gitHostingService match {
-        case MicrositeKeys.GitHub =>
-          s"https://github.com/${gitSettings.githubOwner}/${gitSettings.githubRepo}"
-        case _ => gitSettings.gitHostingUrl
-      }
-    }
-
-    def gitHostingIconClass: String = {
-      gitSettings.gitHostingService match {
-        case MicrositeKeys.GitHub    => "fa-github"
-        case MicrositeKeys.GitLab    => "fa-gitlab"
-        case MicrositeKeys.Bitbucket => "fa-bitbucket"
-        case o: MicrositeKeys.Other  => "fa-git"
-      }
+  def gitSiteUrl: String = {
+    gitSettings.gitHostingService match {
+      case MicrositeKeys.GitHub =>
+        s"https://github.com/${gitSettings.githubOwner}/${gitSettings.githubRepo}"
+      case _ => gitSettings.gitHostingUrl
     }
   }
 
-  case class ExtraMdFileConfig(
-      fileName: String,
-      layout: String,
-      metaProperties: Map[String, String] = Map.empty)
-
-  case class CdnDirectives(jsList: List[String] = Nil, cssList: List[String] = Nil)
-
-  case class ConfigYml(
-      yamlCustomProperties: Map[String, Any] = Map.empty,
-      yamlPath: Option[File] = None,
-      yamlInline: String = ""
-  )
-
-  object ConfigYamlProtocol extends DefaultYamlProtocol {
-
-    implicit object AnyYamlFormat extends YamlFormat[Any] {
-      def write(x: Any): YamlValue = x match {
-        case n: Int            => YamlNumber(n)
-        case n: Long           => YamlNumber(n)
-        case n: Double         => YamlNumber(n)
-        case s: String         => YamlString(s)
-        case b: Boolean        => YamlBoolean(b)
-        case x: Seq[_]         => seqFormat[Any].write(x)
-        case m: Map[String, _] => mapFormat[String, Any].write(m)
-        case t =>
-          serializationError("Serialization Error - Non expected type " + t.getClass.getName)
-      }
-
-      def read(value: YamlValue): Any = value match {
-        case YamlNumber(n)  => n.intValue()
-        case YamlString(s)  => s
-        case YamlBoolean(b) => b
-        case _: YamlArray   => listFormat[Any].read(value)
-        case _: YamlObject  => mapFormat[String, Any].read(value)
-        case x =>
-          deserializationError("Deserialization Error - it failed the deserialization of " + x)
-      }
+  def gitHostingIconClass: String = {
+    gitSettings.gitHostingService match {
+      case MicrositeKeys.GitHub    => "fa-github"
+      case MicrositeKeys.GitLab    => "fa-gitlab"
+      case MicrositeKeys.Bitbucket => "fa-bitbucket"
+      case _                       => "fa-git"
     }
   }
+}
 
+case class ExtraMdFileConfig(
+    fileName: String,
+    layout: String,
+    metaProperties: Map[String, String] = Map.empty)
+
+case class CdnDirectives(jsList: List[String] = Nil, cssList: List[String] = Nil)
+
+case class ConfigYml(
+    yamlCustomProperties: Map[String, Any] = Map.empty,
+    yamlPath: Option[File] = None,
+    yamlInline: String = ""
+)
+
+object ConfigYamlProtocol extends DefaultYamlProtocol {
+
+  implicit object AnyYamlFormat extends YamlFormat[Any] {
+    def write(x: Any): YamlValue = x match {
+      case n: Int            => YamlNumber(n)
+      case n: Long           => YamlNumber(n)
+      case n: Double         => YamlNumber(n)
+      case s: String         => YamlString(s)
+      case b: Boolean        => YamlBoolean(b)
+      case x: Seq[_]         => seqFormat[Any].write(x)
+      case m: Map[String, _] => mapFormat[String, Any].write(m)
+      case t =>
+        serializationError("Serialization Error - Non expected type " + t.getClass.getName)
+    }
+
+    def read(value: YamlValue): Any = value match {
+      case YamlNumber(n)  => n.intValue()
+      case YamlString(s)  => s
+      case YamlBoolean(b) => b
+      case _: YamlArray   => listFormat[Any].read(value)
+      case _: YamlObject  => mapFormat[String, Any].read(value)
+      case x =>
+        deserializationError("Deserialization Error - it failed the deserialization of " + x)
+    }
+  }
 }
